Files updated for module renaming
